Updated EmojiMap; updated tooling to reflect new changes; API cleanup
In this commit, the old "tools" directory was renamed to "cmd" to
reflect idiomatic Go package structures. The extractmap tool was also
updated to reflect new Discord changes.

Along with above updates, mapping.go was regenerated to reflect those
latest changes. It now includes skin tones as well as the latest Discord
emojis.

Example documentations are added into the tests, which should now add a
small "Example" popup in GoDoc.

A function was renamed to GetEntriesWithPrefix with slightly different
behaviors, all of which are documented. This IS a breaking change.